requirements.txt
flaskFlask==3.0.0
pypdf==3.17.0
gunicorn==21.2.0